# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Talent Graph API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients with stats
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ClientListItem" }
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Client detail with roles and placements
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Client detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
        "404": { description: Not found }
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClientUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Client" }
  /placements:
    get:
      operationId: listPlacements
      tags: [placements]
      summary: List placements
      responses:
        "200":
          description: List of placements
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Placement" }
    post:
      operationId: createPlacement
      tags: [placements]
      summary: Create a placement (calculates fee, updates statuses)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlacementInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Placement" }
  /placements/{id}:
    put:
      operationId: updatePlacement
      tags: [placements]
      summary: Update placement status
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlacementUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Placement" }
...
  /clients/{id}/invite:
    post:
      operationId: inviteClientUser
      tags: [clients]
      summary: Invite a user to a client (company) account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InviteClientInput" }
      responses:
        "201":
          description: Invitation sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
...
    InviteClientInput:
      type: object
      required: [name, email]
      properties:
        name: { type: string }
        email: { type: string }
...
    ClientListItem:
      type: object
      required: [id, companyName, industry, status]
      properties:
        id: { type: string }
        companyName: { type: string }
        industry: { type: string }
        size: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        status: { type: string, enum: [ACTIVE, PROSPECT, CHURNED] }
        activeRoles: { type: integer }
        totalPlacements: { type: integer }
        totalFees: { type: number }
    Client:
      type: object
      required: [id, companyName, industry, contactName, contactEmail, status, createdAt]
      properties:
        id: { type: string }
        companyName: { type: string }
        industry: { type: string }
        size: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        contactName: { type: string }
        contactEmail: { type: string }
        contactPhone: { type: ["string", "null"] }
        status: { type: string, enum: [ACTIVE, PROSPECT, CHURNED] }
        roles:
          type: array
          items: { $ref: "#/components/schemas/RoleListItem" }
        placements:
          type: array
          items: { $ref: "#/components/schemas/Placement" }
        totalFees: { type: number }
        createdAt: { type: string }
    ClientInput:
      type: object
      required: [companyName, industry, contactName, contactEmail]
      properties:
        companyName: { type: string }
        industry: { type: string }
        size: { type: ["string", "null"] }
        website: { type: ["string", "null"] }
        logoUrl: { type: ["string", "null"] }
        contactName: { type: string }
        contactEmail: { type: string }
        contactPhone: { type: ["string", "null"] }
        status: { type: string, enum: [ACTIVE, PROSPECT, CHURNED] }
    ClientUpdate:
      allOf:
        - $ref: "#/components/schemas/ClientInput"
    RoleListItem:
      type: object
      required: [id, title, level, department, status, clientName]
      properties:
        id: { type: string }
        title: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        level: { type: string }
        department: { type: string }
        compensationMin: { type: number }
        compensationMax: { type: number }
        feePercentage: { type: number }
        location: { type: ["string", "null"] }
        status: { type: string }
        daysOpen: { type: integer }
        pipelineCount: { type: integer }
        publishedToNetwork: { type: boolean }
...
    Placement:
      type: object
      required: [id, roleId, roleTitle, candidateId, candidateName, clientId, clientName, actualCompensation, feeAmount, status, createdAt]
      properties:
        id: { type: string }
        roleId: { type: string }
        roleTitle: { type: string }
        candidateId: { type: string }
        candidateName: { type: string }
        clientId: { type: string }
        clientName: { type: string }
        actualCompensation: { type: number }
        feePercentage: { type: number }
        feeAmount: { type: number }
        estimatedFee: { type: ["number", "null"] }
        offerDate: { type: ["string", "null"] }
        joinDate: { type: ["string", "null"] }
        status: { type: string, enum: [CONFIRMED, JOINED, PROBATION_PASSED, REPLACEMENT_TRIGGERED] }
        guaranteeExpiry: { type: ["string", "null"] }
        createdAt: { type: string }
    PlacementInput:
      type: object
      required: [roleId, candidateId, actualCompensation]
      properties:
        roleId: { type: string }
        candidateId: { type: string }
        actualCompensation: { type: number }
        offerDate: { type: ["string", "null"] }
        joinDate: { type: ["string", "null"] }
    PlacementUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [CONFIRMED, JOINED, PROBATION_PASSED, REPLACEMENT_TRIGGERED] }
        joinDate: { type: ["string", "null"] }
...